GB - Алгоритмы Урок 1 Задание 1 - с тестами.
